Added cart, Seller and product API

+++ Multi-Vendor/src/main/java/com/vikas/model/CartItem.java
@@ -23,8 +23,8 @@ public class CartItem {
     private String size;
     private int quantity = 1;
 
-    private Integer mrpPrice;
-    private Integer sellingPrice;
+    private double mrpPrice;
+    private double sellingPrice;
     private Long userId;
 
 }
